redistributing tests according to source structure
