--- nmma/tests/data/multi_config_analysis/config.yaml
@@ -1,29 +1,0 @@
-Me2017_1:
-  em-model: Me2017
-  outdir: ${WORKING_DIR}/outdir_1
-  label: injection
-  prior-file: ${WORKING_DIR}/data/multi_config_analysis/Me2017.prior
-  em-tmin: 0.1
-  em-tmax: 5
-  em-tstep: 0.5
-  em-error-budget: 1
-  nlive: 40
-  Ebv-max: 0
-  injection-file: ${WORKING_DIR}/data/multi_config_analysis/injection.json
-  filters: ztfg,ztfr,ztfi
-  remove-nondetections: True
-
-Me2017_2:
-  em-model: Me2017
-  outdir: ${WORKING_DIR}/outdir_2
-  label: injection
-  prior-file: ${WORKING_DIR}/data/multi_config_analysis/Me2017.prior
-  em-tmin: 0.1
-  em-tmax: 5
-  em-tstep: 0.5
-  em-error-budget: 1
-  nlive: 40
-  Ebv-max: 0
-  injection-file: ${WORKING_DIR}/data/multi_config_analysis/injection.json
-  filters: ztfg,ztfr,ztfi
-  remove-nondetections: True
