--- nmma/tests/data/multi_config_analysis/injection.json
@@ -1,31 +1,0 @@
-{
-    "injections": {
-        "__dataframe__": true,
-        "content": {
-            "simulation_id": [
-                0
-            ],
-            "luminosity_distance": [
-                152.52372678074587
-            ],
-            "beta": [
-                3.1674793717549354
-            ],
-            "log10_kappa_r": [
-                1.9033736926068743
-            ],
-            "timeshift": [
-                -1.640734750742769
-            ],
-            "log10_vej": [
-                -1.0471069387482033
-            ],
-            "log10_mej": [
-                -2.9158814967880233
-            ],
-            "geocent_time": [
-                0.09662692256248354
-            ]
-        }
-    }
-}
